backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sendgrid==6.11.0
email-validator==2.1.0
